Add DataShareService for sharing attendance data and update share-state-button functionality
