+++ src/components/share-state-button.ts
@@ -2,10 +2,9 @@ import {html, LitElement} from 'lit'
 import {customElement} from 'lit/decorators.js'
 import {globalStyleSheet} from '../styles/global-style-sheet.ts';
 import {Txt} from '../translations/translations.ts';
-import {globalStore} from '../store/Store.ts';
 import copyImage from '../assets/svg/copy-content.svg'
 import {servicesProvider} from '../services/provider/ServicesProvider.ts';
-import {DataShareService} from '../services/DataShare.service.ts';
+import {ShareUrlService} from '../services/ShareUrl.service.ts';
 
 
 @customElement('share-state-button')
@@ -13,30 +12,8 @@ export class ShareStateButton extends LitElement {
 
 
     async clickOption(): Promise<void> {
-        const state = globalStore.getState();
-        const dataShareService = servicesProvider.getService(DataShareService);
-        const url = dataShareService.getShareUrl(state);
-
-        if (navigator.share) {
-            try {
-                await navigator.share({
-                    title: Txt.shareList,
-                    text: 'Attendance Data',
-                    url: url
-                });
-            } catch (error) {
-                console.error('Error sharing:', error);
-            }
-        } else {
-            console.log('Web Share API not supported', url);
-            try {
-                await navigator.clipboard.writeText(url);
-                alert(Txt.copyContent);
-            } catch (err) {
-                console.error('Failed to copy: ', err);
-            }
-        }
-
+        const shareUrlService = servicesProvider.getService(ShareUrlService);
+        await shareUrlService.shareStateUrl();
     }
